Implement tabbing logic
